+++ CHANGELOG.md
@@ -45,10 +45,12 @@ _Breaking developer changes, which may affect downstream projects or sites that 
 * Fix selection of best background source when starting on a zoomed-out view ([#9325])
 #### :rocket: Presets
 * Render `natural=strait` features in blue color ([#9294])
+* Roads with `motorroad=yes` show implied access restrictions (`foot=no`, `bicycle=no`, `horse=no`) ([id-tagging-schema#609])
 #### :hammer: Development
 
 [#9294]: https://github.com/openstreetmap/iD/issues/9294
 [#9325]: https://github.com/openstreetmap/iD/issues/9325
+[id-tagging-schema#609]: https://github.com/openstreetmap/id-tagging-schema/issues/609
 
 
 # 2.22.0